fix: split Widgetbook catalog parts under 1000 NCL cap (#4332)
Extract Counsel panel and Build port overlay stories into sibling
catalog part files so repo.dart_file_non_comment_line_size passes.
