+++ widgetbook_host/lib/catalogs/catalog.dart
@@ -158,8 +158,10 @@ import 'package:colonizethis_app/widgets/ct_transfer_list.dart';
 // that exceeds the cap is split into clearly-named sibling parts rather than a
 // numbered fragment. The `repo.app_widgetbook_file_naming` gate enforces the
 // no-`catalog_partN` convention.
+part 'catalog_panel_map_build_port_stories.dart';
 part 'catalog_panel_map_stories.dart';
 part 'catalog_panels.dart';
+part 'catalog_panels_counsel.dart';
 part 'catalog_panels_generals.dart';
 part 'catalog_panels_intervention.dart';
 part 'catalog_diplomacy_panel.dart';
