+++ widgetbook_host/lib/catalogs/catalog_panel_map_stories.dart
@@ -121,30 +121,7 @@ List<WidgetbookNode> get provinceOverlayDirectories => [
           buildFortActionHasEngineerUnits: false,
         ),
       ),
-      WidgetbookUseCase(
-        name: 'Standalone — tile Build port enabled',
-        builder: (context) => _provinceOverlayBuildPortStory(
-          showBuildPortActionIcon: true,
-          buildPortActionEnabled: true,
-          buildPortActionHasEngineerUnits: true,
-        ),
-      ),
-      WidgetbookUseCase(
-        name: 'Standalone — tile Build port disabled',
-        builder: (context) => _provinceOverlayBuildPortStory(
-          showBuildPortActionIcon: true,
-          buildPortActionEnabled: false,
-          buildPortActionHasEngineerUnits: false,
-        ),
-      ),
-      WidgetbookUseCase(
-        name: 'Standalone — tile Build port hidden',
-        builder: (context) => _provinceOverlayBuildPortStory(
-          showBuildPortActionIcon: false,
-          buildPortActionEnabled: false,
-          buildPortActionHasEngineerUnits: false,
-        ),
-      ),
+      ...provinceOverlayBuildPortUseCases,
       WidgetbookUseCase(
         name: 'Standalone — tile Purchase land enabled',
         builder: (context) => _provinceOverlayPurchaseLandStory(
@@ -326,33 +303,6 @@ Widget _provinceOverlayBuildFortStory({
       buildFortActionEnabled: buildFortActionEnabled,
       buildFortActionHasEngineerUnits: buildFortActionHasEngineerUnits,
       onBuildFortTap: () {},
-      onClose: () {},
-    ),
-  );
-}
-
-/// MAP20001 Tile **Build port** inline-action variants. Refs #4332.
-Widget _provinceOverlayBuildPortStory({
-  required bool showBuildPortActionIcon,
-  required bool buildPortActionEnabled,
-  required bool buildPortActionHasEngineerUnits,
-}) {
-  final game = demoGameForOverlay;
-  final region = demoRegionForOverlay;
-  return SizedBox(
-    width: 640,
-    height: 520,
-    child: ProvinceSeaZoneDetailOverlay(
-      game: game,
-      region: region,
-      displayId: sampleProvinceIdForOverlay,
-      selectedTileKey: sampleTileKeyForProvinceOverlay,
-      humanPlayerId: game.players.first.id,
-      playerView: demoHumanPlayerViewForOverlay,
-      showBuildPortActionIcon: showBuildPortActionIcon,
-      buildPortActionEnabled: buildPortActionEnabled,
-      buildPortActionHasEngineerUnits: buildPortActionHasEngineerUnits,
-      onBuildPortTap: () {},
       onClose: () {},
     ),
   );
